Fix number display's font size
I changed the font size back to 3 rem because the text in the display
looked too small for comfort. To compensate for the text size, the
error message for dividing by 0 has been shortened to fit the display
with no text wrapping on desktops. For mobile screens, the text still
wraps, but this is a non-issue for the moment.