Remove formatResult function
I have decided not to round the return values of the math functions
or express them as exponents because I simply want to display their
results as is. If users want to work with very large or very small
numbers, they can see their results in their entirety, courtesy of
JavaScript.