Fix functions working with number zero
When a user tries to divide a number by zero, the calculator returned
a string representing an error to the display and then tried to do
math on that string. Unless the user clears the display immediately
after, the calculator would essentially be broken. This is why the
divide function now returns the number entered by the user.

As a result, the enterOperator function no longer checks if there is a
value in the operation array because if any math function returns the
number (not the string) zero, the if statement returns a falsy value
and does not allow the user to perform arithmetic with zero correctly.
The function now depends on how long the array is.

The results array in the enterNumber function does not contain the
string "NONSENSE" because it is unused.