Fix condition in enterOperator
After pressing the equal key, a user could not use the result in the
next calculation because the function checks if the operation array
has a certain length. The calculator would use "null" in the next
calculation with whatever was in the array. Now, the if statement
in the function checks if the first element in the operation array is
empty or not so the user can use a result in their next calculation.